refactor(rust): Move ConsecutiveCountState into support module
