+++ crates/polars-lazy/src/physical_plan/executors/scan/mod.rs
@@ -7,6 +7,8 @@ mod ndjson;
 #[cfg(feature = "parquet")]
 mod parquet;
 
+#[cfg(feature = "ipc")]
+mod support;
 use std::mem;
 #[cfg(any(feature = "parquet", feature = "ipc", feature = "cse"))]
 use std::ops::Deref;
@@ -24,6 +26,8 @@ use polars_io::prelude::*;
 use polars_plan::global::_set_n_rows_for_scan;
 #[cfg(any(feature = "parquet", feature = "csv", feature = "ipc", feature = "cse"))]
 use polars_plan::logical_plan::FileFingerPrint;
+#[cfg(feature = "ipc")]
+pub(crate) use support::ConsecutiveCountState;
 
 use super::*;
 #[cfg(any(feature = "ipc", feature = "parquet"))]
